add decorator pattern
